split maya name tool...
